Tweak Next 13 demo loading (#268)
* Move relay environment to layout.tsx

That way the provider is defined once for the whole app instead of per page.

* Move client Suspense boundaries to after usePreloadedQuery

It seems that using `usePreloadedQuery` inside a `<Suspense>` on a client component will cause it to always suspend on the client, which it doesn't need to. If it does, an hydration error occurs.

Moving to after `usePreloadedQuery`, to the data or fragments that it renders works fine however.

We still need it to be able to render a fallback in case a fragment is refetched.

* Add loading.tsx for equivalent loading status

Since moving the `<Suspense>` boundary down one level in the tree, we no longer get any loading status. Creating a `loading.tsx` wil let us achieve the same effect as before.

This could also be an explicit Suspense boundary around `loadSerializableQuery`.

Removing or renaming `loading.tsx` will let you see the full server-rendered content as a search engine would see it once Next.js releases "Skip streaming for web crawlers" on their [roadmap](https://beta.nextjs.org/docs/app-directory-roadmap).

* Add demo refetch button in Issues

This button is just here to prove showcase that relay has `<Suspense>` to be able refetch data clientside while still rendering a fallback correctly.

I expect to remove this commit later.

* Clarify difference between loading states

Added some extra description text to show if a loading status is from a server component or in the client.

I expect to remove this commit later.

* Move page styles to layout.module.css

Since moving the loading status to loading.tsx from the client suspense, the loading text is no longer positioned like the main content. Moving the styles to the layout resolves this.