Update StoryCommentsSection.tsx to reflect Relay tutorial
Following the steps in the [Connections & Pagination](https://relay.dev/docs/next/tutorial/connections-pagination/#implementing-load-more-comments) part of the Relay Tutorial is clearly states that: 

> At this point, you should see up to three comments on each story. Some stories have more than three comments, and these will show a "Load more" button, although it isn't hooked up yet.

But the example only loads 1 comments, and there is no "Load more" button present. The tutorial assumes that this code is already present and that is confusing to the reader.

In this PR I've updated `StoryCommentsSection.tsx` to reflect the code examples & instructions in the tutorial.

Thanks!